Update svg functionality.

+++ Scripts/Bash/Functions/Svg/svg_updateMetadata.sh
@@ -29,12 +29,14 @@
 
 function svg_updateMetadata {
 
-    local TEMPLATES=''
-    local INSTANCE=''
-    local COUNT=0
     local NAM=''
     local URL=''
     local KEYS=''
+    local FILE=''
+    local COUNT=0
+    local FILES=''
+    local INSTANCE=''
+    local TEMPLATES=''
     local -a TITLE
     local -a VALUE
     local -a PATTERN
@@ -144,10 +146,15 @@ function svg_updateMetadata {
 
     done
 
-    # Define list of files to process.
+    # Redefine filter flag to specify the extension of scalable vector
+    # graphics files we want to update metadata in.  Use action value
+    # as reference to find out different shell files.
+    FLAG_FILTER=".*${FLAG_FILTER}.*\.(svgz|svg)"
+
+    # Build list of files to process.
     cli_getFilesList
 
-    # Process list of files.
+    # Process list of scalable vector graphics.
     for FILE in $FILES;do
 
         # Output action message.
@@ -223,17 +230,13 @@ function svg_updateMetadata {
         sed -i -f $INSTANCE $FILE
 
         # Remove template instance.
-        cli_checkFiles "${INSTANCE}" 'f'
-        rm $INSTANCE
+        if [[ -f $INSTANCE ]];then
+            rm $INSTANCE
+        fi
 
         # Sanitate scalable vector graphic.
         sed -i -r '/^[[:space:]]*$/d' $FILE
 
-    done \
-        | awk -f /home/centos/artwork/trunk/Scripts/Bash/Styles/output_forTwoColumns.awk
-
-    # Check repository changes and ask you to commit them up to
-    # central repository.
-    cli_commitRepoChanges
+    done
 
 }
